First step towards compatibility with VMP 3.8
Not using mnemonics to know if the real stack was written/read.

Also improved code quality.

+++ src/emulator_backend.rs
@@ -29,7 +29,7 @@ fn get_emulator(
 
     // loading the segments
     for element in data {
-        uc.mem_map(element.start, element.size as usize, Permission::all())?;
+        uc.mem_map(element.start, element.size, Permission::all())?;
         uc.mem_write(element.start, &element.data)?;
     }
